revert: Rollback extracting Months and Years components
